restructuring tools
